feat(flights-widget): add flight fetcher DOT-309
