+++ .babelrc
@@ -1,4 +1,12 @@
 {
-  "presets": ["es2015", "react", "stage-1"],
-  "plugins": ["transform-decorators-legacy", "transform-object-assign"]
+  "presets": [
+    "es2015",
+    "react",
+    "stage-1"
+  ],
+  "plugins": [
+    "transform-decorators-legacy",
+    "transform-object-assign",
+    "transform-runtime"
+  ]
 }
