+++ package.json
@@ -39,6 +39,7 @@
   },
   "dependencies": {
     "axios": "^0.15.3",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "chai": "^3.5.0",
     "classnames": "^2.2.5",
     "clipboard": "^1.7.1",
@@ -46,6 +47,7 @@
     "fuzzy": "^0.1.3",
     "haversine": "^1.0.1",
     "jquery": "^3.1.1",
+    "js-cookie": "^2.2.0",
     "leaflet": "^1.0.1",
     "lodash": "^4.16.6",
     "moment": "^2.20.1",
